// frontend/src/api/generated/model/positionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * crossbill API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Position in a book document.
 */
export interface PositionResponse {
  /** Document-order element number */
  index: number;
  /** Character offset within element */
  char_index: number;
}